requirements.txt
Django==1.6.5
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
djangorestframework==2.3.14
enum34==1.0
google-api-python-client==1.2
gunicorn==19.1.0
httplib2==0.9
psycopg2==2.5.3
python-gflags==2.0
static3==0.5.1
urllib3==1.9
wsgiref==0.1.2
